refactor: rename job classes and update package imports for consistency
