+++ bottlenote-batch/src/main/java/app/batch/bottlenote/schedule/BatchSchedules.java.bak
@@ -9,7 +9,7 @@ import org.springframework.batch.core.launch.JobLauncher;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import static app.batch.bottlenote.job.BestReviewSelectionJob.BEST_REVIEW_JOB_NAME;
+import static app.batch.bottlenote.job.BestReviewSelectionJobConfig.BEST_REVIEW_JOB_NAME;
 import static app.batch.bottlenote.job.PopularAlcoholSelectionJob.POPULAR_JOB_NAME;
 import static java.time.LocalDateTime.now;
