+++ bottlenote-batch/src/main/java/app/batch/bottlenote/schedule/BestReviewQuartzJob.java
@@ -1,11 +1,11 @@
 package app.batch.bottlenote.schedule;
 
 
+import static app.batch.bottlenote.job.BestReviewSelectionJobConfig.BEST_REVIEW_JOB_NAME;
+
 import org.springframework.batch.core.configuration.JobRegistry;
 import org.springframework.batch.core.launch.JobLauncher;
 import org.springframework.stereotype.Component;
-
-import static app.batch.bottlenote.job.BestReviewSelectionJob.BEST_REVIEW_JOB_NAME;
 
 /**
  * 베스트 리뷰 선정을 위한 Quartz Job 구현 클래스입니다.
